Control.Arrow.Tree

Documentation

newtype ATree t a b

Like Tree, but the arguments are applied differently, to allow it to be an Arrow

Constructors

ATree

Fields

getATree :: Tree (t a) b

newtype AForest t a b

See AForest

Constructors

AForest

Fields

getAForest :: Forest (t a) b

comp :: Tree ((->) b) c -> Tree ((->) a) b -> Tree ((->) a) c

Compose two Trees

getTree2 :: (Tree ma a -> Tree mb b -> Tree mc c) -> (a, Forest ma a) -> (b, Forest mb b) -> (c, Forest mc c)

coerce

arrCoerce :: (Arrow a, Coercible c d) => a b c -> a b d

unsafeCoerce is used for now, since the following error appears:

 • Couldn't match representation of type ‘t a (c, Forest (t a) c)’
                            with that of ‘t a (Tree (t a) c)’
     arising from a use of ‘coerce’
   NB: We cannot know what roles the parameters to ‘t a’ have;
     we must assume that the role is nominal

This is safe by the following:

 (arr coerce .)   :: (Arrow t, Coercible b c) => t a b -> t a c -- initial definition
   = (coerce .)   :: (Arrow t, Coercible b c) => t a b -> t a c -- arr id = id
   = coerce       :: (Arrow t, Coercible b c) => t a b -> t a c -- (id .) = id

goal' :: Arrow t => ATree t b c -> ATree t a b -> ATree t a c

coerce more strangely doesn't work here either. weird

goal :: Arrow t => Tree (t b) c -> Tree (t a) b -> Tree (t a) c

Goal function

trans1 :: ((a, Forest ma a) -> (b, Forest mb b) -> (c, Forest mc c)) -> Tree ma a -> Tree mb b -> Tree mc c

coerce

goal1 :: Arrow t => (c, Forest (t b) c) -> (b, Forest (t a) b) -> (c, Forest (t a) c)

trans2 goal2

trans2 :: Arrow t => ((c, t b (Tree (t b) c)) -> (b, t a (Tree (t a) b)) -> (c, t a (Tree (t a) c))) -> (c, Forest (t b) c) -> (b, Forest (t a) b) -> (c, Forest (t a) c)

coerce

goal2 :: Arrow t => (c, t b (Tree (t b) c)) -> (b, t a (Tree (t a) b)) -> (c, t a (Tree (t a) c))

Second goal function

 goal2 (x, fx) (y, fy) = (x, arr (uncurry goal) . (fx . arr (const y) &&& fy))

laog' :: Arrow t => AForest t b c -> AForest t a b -> AForest t a c

Coerce doesn't work here, which is too bad

laog :: Arrow t => Forest (t b) c -> Forest (t a) b -> Forest (t a) c

snart1 laog1

snart1 :: Arrow t => (t b (Tree (t b) c) -> t a (Tree (t a) b) -> t a (Tree (t a) c)) -> Forest (t b) c -> Forest (t a) b -> Forest (t a) c

coerce

laog1 :: Arrow t => t b (Tree (t b) c) -> t a (Tree (t a) b) -> t a (Tree (t a) c)

snart2 laog2

snart2 :: Arrow t => (t b (c, Forest (t b) c) -> t a (b, Forest (t a) b) -> t a (c, Forest (t a) c)) -> t b (Tree (t b) c) -> t a (Tree (t a) b) -> t a (Tree (t a) c)

By the logic of arrCoerce, this use of unsafeCoerce is safe

laog2 :: Arrow t => t b (c, Forest (t b) c) -> t a (b, Forest (t a) b) -> t a (c, Forest (t a) c)

Opposite of goal2

idForest' :: Arrow t => AForest t a a

Wow, coerce doesn't work here. I had to use the AForest constructor instead

Notes:

So, technically, it appears that Tree is only a category when we have a left-identity for `Arrow a`

We can however do stuff like:

 a -> Tree (t a) a
 t a (Tree (t a) a

idTree :: Arrow t => Tree (t a) a
idTree1 :: Arrow t => (a, Forest (t a) a)

idForest :: Arrow t => Forest (t a) a

See idForest

transIdForest1 :: Arrow t => t a (Tree (t a) a) -> Forest (t a) a

coerce

idForest1 :: Arrow t => t a (Tree (t a) a)

transIdForest2 idForest2

transIdForest2 :: Arrow t => t a (a, Forest (t a) a) -> t a (Tree (t a) a)

arrCoerce

idForest2 :: Arrow t => t a (a, Forest (t a) a)

Coerced all the way up to idForest

forestInv :: (Arrow t, Groupoid t) => AForest t a b -> AForest t b a

undefined

forestInv1 :: (Arrow t, Groupoid t) => Forest (t a) b -> Forest (t b) a

undefined

forestInv2 :: (Arrow t, Groupoid t) => t a (Tree (t a) b) -> t b (Tree (t b) a)

undefined

swaps :: Arrow t => t a (Tree (t a) b) -> t (Tree (t b) a) b

undefined

swaps1 :: Arrow t => t a (b, Forest (t a) b) -> t (a, Forest (t b) a) b

undefined

swaps2 :: Arrow t => t a (b, t a (Tree (t a) b)) -> t (a, t b (Tree (t b) a)) b

undefined. Ok, I'm convinced that this needs Groupoid

bifoldableNotes :: ()

Notes on Data.Bifoldable, Data.Bitraversable, etc:

Data.Bifoldable
instance Bifoldable t => Bifoldable (ATree t) where
  bifoldr :: (a -> c -> c) -> (b -> c -> c) -> c -> t a b -> c
  bifoldr :: (a -> c -> c) -> (b -> c -> c) -> c -> ATree t a b -> c
  bifoldr :: (a -> c -> c) -> (b -> c -> c) -> c -> Tree (t a) b -> c

This makes me think that Forest (t a) b ~ Forest (Forest t a) b, which would be amazeballs
Maybe something similar, like the wanted functions?

instance Bifoldable t => Bifoldable (AForest t) where
  bifoldr :: (a -> c -> c) -> (b -> c -> c) -> c -> t a b -> c
  bifoldr :: (a -> c -> c) -> (b -> c -> c) -> c -> Forest (t a) b -> c

Data.Bitraversable
class (Bifunctor t, Bifoldable t) => Bitraversable t
  bitraverse :: Applicative f => (a -> f c) -> (b -> f d) -> t a b -> f (t c d)

Data.Bifunctor.Functor
type (:->) p q = forall a b. p a b -> q a b

class BifunctorFunctor t where
  bifmap :: (p :-> q) -> t p :-> t q

wantTo :: AForest t a (ATree t a b) -> AForest t (ATree t a b) b

undefined

wantFrom :: AForest t (ATree t a b) b -> AForest t a (ATree t a b)

undefined

sndIn :: ((a, b), c) -> ((a, c), b)

Re-associate a tuple

assoc :: ((a, b), c) -> (a, (b, c))

I'm pretty surprised that this isn't in base

unassoc :: (a, (b, c)) -> ((a, b), c)

Inverse of assoc

treeArr :: Arrow a => Tree ((->) b) c -> Tree (a b) c

Lift functions, inside of a Tree, into Arrows

Scratch notes:

arr :: Arrow a => (b -> c) -> a b c
arr . const :: Arrow a => c -> a b c

arr Tree    :: Arrow a => a (c, Forest (a b) c) (Tree (a b) c)
arr getTree :: Arrow a => a (Tree (a b) c) (c, Forest (a b) c)

arr Forest    :: Arrow a => a (a b (Tree (a b) c)) (Forest (a b) c)
arr getForest :: Arrow a => a (Forest (a b) c) (a b (Tree (a b) c))

Wow, this just blew my mind. If we have cat be an arrow, then the pair operations that I use above are already given!!

TreeArrow (Tree (f, fs)) . TreeArrow (Tree (g, gs)) = TreeArrow (Tree (f, fs . gs))

compf :: Forest ((->) b) c -> Forest ((->) a) b -> Forest ((->) a) c
compf :: (b -> Tree ((->) b) c) -> (a -> Tree ((->) a) b) -> (c -> Tree ((->) a) c)
compf :: (b -> (c, Forest ((->) b) c)) -> (a -> (b, Forest ((->) a) b)) -> (c -> (c, Forest ((->) a) c))

I'm willing to bet that Tree is analogous to True (as it's inhabited) and Forest is analogous to False (as it's not, yet):

  Tree && Tree forms a Tree:   (a & b -> c) -> (Tree m a & Tree m b) -> Tree m c
  Tree || Forest forms a Tree: (a | b -> c) -> (Tree m a | Forest m b) -> Tree m c
  Forest || Forest forms a Tree: (a | b -> c) -> (Forest m a | Forest m b) -> Forest m c

Ohhh, can we also lift inside? Duh, as long as m is applicative!

  (a | b -> c) -> Tree m (a | b) -> Tree m c

Can we convert a comonad to a monad?

  (w a -> a) looks like a monad to me, in fact, it appears to be: Forest (CoKleisi a) b,
    Cokleisli { runCokleisli :: w a -> b }
    Instances: .. Monad (Cokleisli w a)

Can we convert a monad to a comonad?

  (a, a -> m a) looks like a comonad to me, in fact it appears to be: Tree State a

forestArr :: Arrow a => Forest ((->) b) c -> Forest (a b) c

Forest counterpart to treeArr