Control.Arrow.Tree.Experimental

Documentation

forestInv :: (Arrow t, Groupoid t) => AForest t a b -> AForest t b a

undefined

forestInv1 :: (Arrow t, Groupoid t) => Forest (t a) b -> Forest (t b) a

undefined

forestInv2 :: (Arrow t, Groupoid t) => t a (Tree (t a) b) -> t b (Tree (t b) a)

undefined

swaps :: Arrow t => t a (Tree (t a) b) -> t (Tree (t b) a) b

undefined

swaps1 :: Arrow t => t a (b, Forest (t a) b) -> t (a, Forest (t b) a) b

undefined

swaps2 :: Arrow t => t a (b, t a (Tree (t a) b)) -> t (a, t b (Tree (t b) a)) b

undefined. Ok, I'm convinced that this needs Groupoid

wantTo :: AForest t a (ATree t a b) -> AForest t (ATree t a b) b

undefined

wantFrom :: AForest t (ATree t a b) b -> AForest t a (ATree t a b)

undefined

Orphan instances

Biapplicative t => Biapplicative (AForest t)
This also looks a lot like distribution.. (undefined)

Methods

bipure :: a -> b -> AForest t a b

(<<*>>) :: AForest t (a -> b) (c -> d) -> AForest t a c -> AForest t b d

(*>>) :: AForest t a b -> AForest t c d -> AForest t c d

(<<*) :: AForest t a b -> AForest t c d -> AForest t a b

Biapplicative t => Biapplicative (ATree t)

Methods

bipure :: a -> b -> ATree t a b

(<<*>>) :: ATree t (a -> b) (c -> d) -> ATree t a c -> ATree t b d

(*>>) :: ATree t a b -> ATree t c d -> ATree t c d

(<<*) :: ATree t a b -> ATree t c d -> ATree t a b

Choice t => Choice (AForest t)
This is harder than I thought, might be impossible, probably impossible.

Try extending Choice (ATree t) to not use Choice (AForest t).

(Currently, both are undefined.)

Methods

left' :: AForest t a b -> AForest t (Either a c) (Either b c)

right' :: AForest t a b -> AForest t (Either c a) (Either c b)

Choice t => Choice (ATree t)

Methods

left' :: ATree t a b -> ATree t (Either a c) (Either b c)

right' :: ATree t a b -> ATree t (Either c a) (Either c b)

Closed t => Closed (AForest t)
REQUIRES DISTRIBUTION: distribute :: Functor f => f (g a) -> g (f a), Functor f => f (Tree m a) -> Tree m (f a)

Methods

closed :: AForest t a b -> AForest t (x -> a) (x -> b)

Closed t => Closed (ATree t)

Methods

closed :: ATree t a b -> ATree t (x -> a) (x -> b)

(Arrow t, Groupoid * t) => Groupoid * (AForest t)
inv = forestInv

Methods

inv :: k1 a b -> k1 b a

(Arrow t, Groupoid * t, TypeError Constraint (Text "Impossible instance, no way to convert @a@ to @b@.")) => Groupoid * (ATree t)
Impossible instance, no way to convert a to b.

Methods

inv :: k1 a b -> k1 b a