Control.Monad.Trans.TreeT

Documentation

algebraicNotes :: ()

Algebraic notes:

-- f' = (f / g) * g'
-- f' = (f  g) * ((g  f) * f')
-- f' = (*) (() f g) ((*) (() g f) f')
-- f' = (f y g) x ((g y f) x f')

f' x y f g = x (y f g) . x (y g f) $ f'

-- g' = (g / f) * f'
-- g' = (g  f) * ((f  g) * g')
-- g' = (g y f) x ((f y g) x g')

g' x y f g = x (y g f) . x (y f g) $ g'

f' = (f y g) x g'
g' = (g y h) x h'
h' = (h y f) x f'

If x right-distributes over y:

 f' = (f y g) x g' = (f y g') x (g y g')
 g' = (g y h) x h' = (g y h') x (h y h')
 h' = (h y f) x f' = (h y f') x (f y f')

 f' = (f y g) x g' = (f y g) x ((g y h) x ((h y f) x f'))
 g' = (g y h) x h' = (g y h) x ((h y f) x ((f y g) x g'))
 h' = (h y f) x f' = (h y f) x ((f y g) x ((g y h) x h'))

 f' = x (y f g) . x (y g f) $ f'
 g' = x (y g f) . x (y f g) $ g'

 f' = x (y f g) g' = x (y f g) . x (y g h) . x (y h f) $ f'
 g' = x (y g h) h' = x (y g h) . x (y h f) . x (y f g) $ g'
 h' = x (y h f) f' = x (y h f) . x (y f g) . x (y g h) $ h'

 f' = x (y f g) g' = ((.) (x (y f g)) ((.) (x (y g h)) (x (y h f))) $ f'
 g' = x (y g h) h' = x (y g h) . (x (y h f) . x (y f g)) $ g'
 h' = x (y h f) f' = x (y h f) . (x (y f g) . x (y g h)) $ h'

 (.) :: cat b c -> cat a b -> cat a c

 (.) (x (y g h)) (x (y h f)) :: xygh b1 c1 -> xyhf a1 b1 -> ghf a1 c1
 (.) (x (y h f)) (x (y f g)) :: xyhf b2 c2 -> xyfg a2 b2 -> hfg a2 c2
 (.) (x (y f g)) (x (y g h)) :: xyfg b3 c3 -> xygh a3 b3 -> fgh a3 c3

 (.) (x (y f g)) :: xyfg c1 d1 -> ghf a1 c1 -> C1 a1 d1
 (.) (x (y g h)) :: xygh c2 d2 -> hfg a2 c2 -> C2 a2 d2
 (.) (x (y h f)) :: xyhf c3 d3 -> fgh a3 c3 -> C2 a3 d3

So x has to be at least a semigroupoid, to support composition

   furthermore, it has to lift all the pairs to a common category
   furthermore, it has to be funtorial, to allow preservation of internal properties
 y :: `f -> `g -> `(y f g)
 y :: `g -> `f -> `(y g f)

 y :: `f -> `g -> `(y f g)
 y :: `g -> `h -> `(y g h)
 y :: `h -> `f -> `(y h f)

algebraicExamples :: ()

Algebraic examples:

 La * Lb -> La

 (La :*: Lb) a -> La a

 (La -> (La || Lb)) -> (Lb -> (La || Lb)) -> (La -> La - Lb)

 Suppose La <= Lb

 A, B, A && B, A || B, A * B, A + B

 Either A B
 (A, B)

considerLanguageCycle :: ()

propertyPreservation :: ()

Thus we see that for languages based on (->), aka application, every property seems to hold on either Tree or Forest:

Because Tree is the combination of the Value language and the Action language

Because ATree is the combination of the Semigroupoid language with the Category language (ATree doesn't have identity, AForest does, even though it contains ATree.)

OR because ATree is the product of (Category - Identity) and Category..?

So all associtive-requiring properties hold for both and all associative+identity-requiring properties only hold for AForest

If this construction works, then I believe it may be "self-holding", i.e. free: The primitives used to express it are exactly the primitives needed to express it.

If this works and forms an internally consistent language, then this effectively adds a primitive form of implementation-agnostic, decidable, recursion to Flock.

It also allows the simple addition of Categories, Application, Type hierarchies, etc. etc. etc.

Note: While I'm unsure whether any n-lang in the Flock core is sufficient to express itself, it's trivial to show that there's an (n+m)-lang that does (because of regularity).

Does this preserve: consistency, completeness, decidability, finiteness, haltingness, etc?

Well, if it's shown on a nice category-theoretic level, it doesn't really matter, no?

 Everything will be preserved, [evil-laugh] XD

(That still cracks me up)

Natural extension of A. There is a function

 f : A -> A',
 f' :: A' -> A,
 f' . f = id,
 A' ~ A = f . f' = id

I wonder if this indeed forms a structure analogous to a factorization system for arbitrary categories?

twoMoreAttempts :: ()

More possibilities

f :: a -> a',
f' :: a' -> a,
f' . f = id,
Iso (a' ~ a) (f . f' = id)

naturalTransformationAlgebra :: ()

Summarize and notes:

F, G functors from C to D
n, natural transformation F -> G
  nx :: forall (x :: C). F x -> G x
  forall (f :: (x :: C) -> (y :: C)).
    exists (ny :: F y -> G y).
    ny . F (f :: x -> y) == G (f :: x -> y) . nx

Functor (f :: * -> *)
Functor (g :: * -> *)

f' = (f / g) * g'
f' = (f  g) * ((g  f) * f')
f' = (*) (() f g) ((*) (() g f) f')
f' = (f y g) x ((g y f) x f')
f' = x (y f g) (x (y g f) f')

g' = (g / f) * f'
g' = (g  f) * ((f  g) * g')
g' = (g y f) x ((f y g) x g')

ttDefinition :: ()

Free/Cofree and algebra

newtype Free f a = Free { unfree :: Either a (f (Free f a))) }
  Either a (f (Free f a))
  (Either a :.: f :.: Free f)
Sum version of Tree f a

Cofree f a =  a :<  (f (Cofree f a))
  Tree f a =  a (,) (f (Tree   f a))

Free f a = Pure a | Free (f (Free f a))
Free f a = Either a (f (Free f a))
Free f a = Either a (f (Either a (f (Free f a))))
Free f a = Either a (f -> (a, (Either a (f -> (a, )(Free f a)))))
Free (TT f a b) c = Either a (TT a b (Free (TT f a b) c))
Free (TT f a b) c = Either a (TT a b (Either a (TT a b (Free (TT f a b) c))))

Forest (Either a . f) a = Either a (f (a, Forest f a))
Forest f a = f (a, Forest f a)
Forest f a = f (Tree f a)

TT f a b = f a -> (a, b)

freeAndCofree :: ()

Free and Cofree instances:

newtype FreeT   f m a = FreeT   { runFreeT   :: m (FreeF f a (FreeT f m a)) }
  newtype FreeT   f m a = m (Either a (f (FreeT f m a)))
  newtype FreeT   f m a = m (Either a (f (m (Either a (f (FreeT f m a))))))
data FreeF   f a b = Pure a | Free (f b)
  data FreeF   f a b = Either a (f b)

newtype CofreeT f w a = CofreeT { runCofreeT :: w (CofreeF f a (CofreeT f w a)) }
  newtype CofreeT f w a = w (a, f (CofreeT f w a))
  newtype CofreeT f w a = w (a, f (w (a, f (CofreeT f w a))))
data CofreeF f a b = a :< (f b)
  data CofreeF f a b = (a, f b)

Functor f => Monad (Free f a) => MonadFree f (Free f) =>
  wrap :: f (Free f a) -> Free f a
  wrap x = Pure  (x :: f (Free f a))
  wrap x = Right (x :: f (Free f a))

Monad f => Monad (Forest f a) =>
  wrap :: f (Forest f a) -> Forest f a
  wrap = (coerce :: (f (f (a, Forest f a)) -> f (a, Forest f a)) -> f (Forest f a) -> Forest f a) (join :: Monad f => f (f (a, Forest f a)) -> f (a, Forest f a))

newtype Tree m a

A Generalized tree type

Question: does any property such that

  Prop :: * -> Constraint
    Prop a
      => Prop (Tree Identity a)?
    Prop a, Prop a + Prop b => Prop (a, b), Prop a => Prop (m a)
      => Prop (Tree m a)?
    Prop a => Prop (m a)
      => Prop (Tree m a)?
  Prop :: (* -> *) -> Constraint
    Prop Identity, Prop a + Prop b => Prop (a :*: b), Prop m
      => Prop (Tree m)?
    Prop m
      => Prop (Forest m)?
  Prop :: (* -> * -> *) -> Constraint
    Prop (,), Prop t
      => Prop (ATree t m)?
    Prop t
      => Prop (AForest t m)?

E.g.

 instance Alternative m => Alternative (Forest m)@

But no instance for Tree m

E.g.

 instance Functor m => Comonad (Tree m)

But no instance for Forest m

Constructors

Tree

Fields

getTree :: (a, Forest m a)

newtype Forest m a

The counterpart to Tree

Constructors

Forest

Fields

getForest :: m (Tree m a)

liftTree :: Monad m => m a -> m (Tree m a)

Tree is not a monad transformer, but Tree . m is, for Monad m

ideaInstanceEnum :: ()

Lift an Enum to a Tree type where the outermost is first?

Hmm.. I know that we can't use the internal monadic structure of Tree since we know nothing about it.

What would we like to do with this? well, it might be nice to add labeling to a Tree using Enum

 instance Enum a => Enum (Tree m a) where

 -- The functions below are for Forest, since we probably can't do fromEnum

 -- | the successor of a value. For numeric types, succ adds 1.
 succM :: a -> a

 -- | the predecessor of a value. For numeric types, pred subtracts 1.
 predM :: a -> a

 -- | Convert from an Int.
 toEnumM :: Int -> a

 -- | Convert to an Int. It is implementation-dependent what fromEnum returns when applied to a value that is too large to fit in an Int.
 fromEnumM :: a -> Int

 -- | Used in Haskell's translation of [n..].
 enumFromM :: a -> [a]

 -- | Used in Haskell's translation of [n,n'..].
 enumFromThenM :: a -> a -> [a]

 -- | Used in Haskell's translation of [n..m].
 enumFromToM :: a -> a -> [a]

 -- | Used in Haskell's translation of [n,n'..m].
 enumFromThenToM :: a -> a -> a -> [a]

pullM :: Functor m => Forest m a -> m (Forest m a)

Pull out a layer of monadic context from a Forest.

Not sure, but I think that the only inhabitants of this type can:

return the original

do this

swap two Tree values

some combination of the above

pullMa :: Functor m => Forest m a -> m a

pullM and extract

unfoldTree :: Functor m => (b -> (a, m b)) -> b -> Tree m a

Unfold a Tree from a seed value

unfoldForest :: Functor m => (b -> m (a, b)) -> b -> Forest m a

Unfold a Forest from a seed value

iterateTree :: Applicative m => (a -> m a) -> a -> m (Tree m a)

iterateForest :: Applicative m => (a -> m a) -> a -> Forest m a

cmpLength :: [a] -> [b] -> Ordering

Compare length of lists, without Int or Eq

minsBy :: (a -> a -> Ordering) -> [a] -> [a]

Minimums by: