Data.Bidistributable

Documentation

class Bifunctor p => Bidistributable p f where

Distribute over a Bifunctor

Minimal complete definition

bidistribute

Methods

bidistribute :: f (p a b) -> f (p (f a) (f b))

Later called "push" and or "push2"

class Swap p where

Minimal complete definition

swap

Methods

swap :: p a b -> p b a

newtype RopeT p m a b

Rope transformer type

Constructors

RopeT

Fields

pullRopeT :: p a (m (RopeT p m b a))

commonVariableTypes :: ()

Commonly used variable types:

f :: t -> t
g :: t -> t
a :: t

functorNotes :: ()

Let C be a category. F associates (X is an object in C and G(X) is as well):

  a  :: X      => F(a) :: F(X)
  f  :: X -> Y => F(f) :: F(X) -> F(Y)
  id_X :: X -> X => F(id_X) = id_F(X) :: F(X) -> F(X)
  f :: X -> Y, g :: Y -> Z, g . f :: X -> Z => F (g . f) = F(g) . F(f) :: F(X) -> F(Z)

                     G associates (X is an object in C and G(X) is as well)
  a  :: X      => G(a) :: G(X)
  f  :: X -> Y => G(f) :: G(X) -> G(Y)
  id_X :: X -> X => G(id_X) = id_G(X) :: G(X) -> G(X)
  f :: X -> Y, g :: Y -> Z, g . f :: X -> Z => G (g . f) = G(g) . G(f) :: G(X) -> G(Z)

functorIsomorphismCompositions :: ()

We want: functor isomorphisms between these compositions of F and G:

  F(G(A))    -- we have a composition of functors
  F(F(G(A))) -- we partition the outer functor into pieces such that `unique partition (xs :: f (g a))` is true for our partition (if it exists, of course, that's where graph coloring comes in)
  F(G(F(A))) -- we push the partition into the inner functor, resulting in the connected (f (g a)) pieces being joined along their partitions.

In other words, it pulls the partition up to F's level while pushing the structure of F down into G. The structure of F is available to G in slices (local by definition of the partition).

Note: all of these operations are invertible since we have unique partitions.

Additionally, the functors which this works for are local functors and computers _really_ like computational and data locality. Like they simply adore it. Computational locality makes stream and massively parallel processing a breeze (case in point, the line-by-line parser I just wrote)

naturalIsomorphisms :: ()

That is, we want three natural transformations that are also isomorphisms between those compositions.

  f :: X -> Y => F(G(f)) :: F(G(X)) -> F(G(Y))

  X :: C => eta_1(X) :: F(G(X)) -> F(F(G(X))) :: C -> C
  X :: C => eta_2(X) :: F(F(G(X))) -> F(G(F(X))) :: C -> C
  X :: C => eta_3(X) :: F(G(F(X))) -> F(G(X)) :: C -> C

  f :: X -> Y :: C -> C => eta_1(Y) . F(G(f)) = F(F(G(f))) . eta_1(X) :: F(G(X)) -> F(F(G(Y))) :: C -> C
  f :: X -> Y :: C -> C => eta_2(Y) . F(F(G(f))) = F(G(F(f))) . eta_2(X) :: F(F(G(X))) -> F(G(F(Y))) :: C -> C
  f :: X -> Y :: C -> C => eta_3(Y) . F(G(F(f))) = F(G(f)) . eta_3(X) :: F(G(F(X))) -> F(G(Y)) :: C -> C

We require these natural transformations to be bijections.

The three base words (those functor compositions) are equivalent
Functor composition is associative
We have a compositional identity, the identity functor

regularLanguage :: ()

We thus have a group with the binary operation (.), identity the category identity, every element is invertible because we have isomorphisms between the compositions.

It has the presentation: f g = f f g = f g f.

Since it is a group, we can form the language of elements equivalent to those three words.

This diagram makes it clear:

(the lines are equivalence and the . is composition)

            (f . g)
              / 
             /   
            /     
           /       
          /         
  f . (f . g) --- (f . g) . f

First of all, we have the word "fg".

  Next, we have the word formed by performing the substitution "fg -> fgf", namely "fgf".
  Next, we have the word formed by performing the substitution "fg -> ffg", namely "ffg".

The substitution rules in full are as follows (with the substring "fg" wrapped in parentheses for emphasis alone):

 "(fg)  -> f(fg)"
 "f(fg) -> (fg)"
 "(fg)  -> (fg)f"
 "(fg)f -> (fg)"
 "f(fg) -> (fg)f"
 "(fg)f -> f(fg)"
 ----------------
   It's a rote exercise to show that any rule of the form "f?(fg)f? -> f?(fg)f?" can be derived.

Since the substring "fg" can never be broken by the available base rules, we simply have the ability to induct on "f?(fg)f? -> f?(fg)f?" to form "f{n}(fg)f{m}" for any n,m.

We simply compose the base rule with itself (max n m) times. We then resolve the optionals (?'s) to the number of "f"'s desired on either side of the "(fg)" center.

We thus have that this language is exactly the regular language: "(f*)(fg)(f*)"

exampleApplications :: ()

One example application is [], Maybe, where pushing ([Maybe a] -> [Maybe [a]]) partitions by isJust. This is stream-friendly.

Another example is ListT IO, Threaded, where Threaded is a comonad that evaluates its contents on extract, which we can do purely since we're still inside of ListT IO and couldn't do otherwise.

Pushing in is some parallel execution strategy, in that case.

In other words, we can use this to treat monads as comonads. We just do: (Monad m => Turn (Compose m f) m) and the inner m can be treated as a comonad inside the turn.

I.e. we have: Monad m => m (Comonad m => ())

This could be well-typed if GHC supported impredicative polymorphism, but it doesn't.

How does it work?

 Well, inside of m, we have extract (m (m a) -> m a), which is just join
 Also, inside of m, we still have return so then we get duplicate (m (m a) -> m (m (m a))), which is return = fmap return

Inside of a comonad, we get a monad. This is the free comonad (/monad)

Ahhh, don't forget about orthogonal lists, e.g. a 2d array. this may be able to pass layers in and out

effectiveIdempotencePush :: ()

Really, we want two applications of push to be equivalent to pure push (we don't gain anything from pushing more) (the benefit is that it's equivalent to the limit of the other pushes OR you can consider it as a guarantee that everything to be pushed is pushed in the first push)

This is: "pushes are effectively idempotent" (using some "generic" injection into f to show that additional pushes do no more than the generic injection)

  push  :: forall (a :: c). f (g a) -> f (g (f a))
  puref :: forall (a :: c). a -> f a

  push      . push :: forall (a :: c). f (g a) -> f (g (f (f a)))
  fmap (fmap puref) . push :: forall (a :: c). f (g a) -> f (g (f (f a)))
  push . push == fmap (fmap puref) . push

This is: "pushes are effectively idempotent" (using some "generic" extraction from f (which arguably, we can guarantee in some way when we want all of the strcuture of f that can be pushed to be))

  push     :: forall (a :: c). f (g a) -> f (g (f a))
  extractf :: forall (a :: c). f a -> a

  push . push                :: forall (a :: c). f (g a) -> f (g (f (f a)))
  fmap (fmap extractf) . push . push :: f (g a) -> f (g (f a))

  fmap (fmap extractf) . push . push == push

This is: "pushes are effectively idempotent" (using some "generic" joining of fs)

  push  :: forall (a :: c). f (g a) -> f (g (f a))
  joinf :: forall (a :: c). f (f a) -> f a

  push . push                :: forall (a :: c). f (g a) -> f (g (f (f a)))
  fmap (fmap joinf) . push . push :: f (g a) -> f (g (f a))

  fmap (fmap joinf) . push . push == push

This is: "pushes are effectively idempotent" (using some "generic" duplication of f's)

  push       :: forall (a :: c). f (g a) -> f (g (f a))
  duplicatef :: forall (a :: c). f a -> f (f a)

  push . push                :: forall (a :: c). f (g a) -> f (g (f (f a)))
  fmap (fmap duplicatef) . push      :: forall (a :: c). f (g a) -> f (g (f (f a)))

  fmap (fmap duplicatef) . push == push . push

weakestVersionDuplicate :: ()

I believe that the weakest version is a duplicate that is injective?

Hmm, suppose we are within a context. This context is composable with itself.

We want any - 1, which is equivalent to having 2 - 1.

We want this to be a natural transformation.

Ahh, what if we had: f (f a) -> Maybe (f a)

Then, we could define the other mappings to match it so that it's always Just.

Regardless, we want to be able to show that the additional compositions provided by pushes are trivial.

leftPartitionProperty :: ()

Really, we probably want to have this property on the left as well, namely: partitioning an already partitioned functor only adds a trivial layer, since the functor has already been partitioned.

  partition :: f (g a)  -> f (f (g a))
  fmap partition . partition = fmap puref . partition
  fmap extract . fmap partition . partition = partition

leastActionExhaustivity :: ()

leastLeftAction :: ()

linearizationQuestion :: ()

Question: can we linearize this?

Yes, we just map words of functions of type f (g a) -> f (f (g a)) using the following patterns:

 f0
 fmap f0 . f1
 fmap (fmap f0) . fmap f1 . f2

Quality preserves the lengths of words, because the words have unique types.

We want that (p=partition):

"pp" -> "_p"

Now, what is this blank? Hmm.

 eta(x) :: f (g (f x)) -> f (g (f (f x)))
 fn :: x -> y => eta(y) . f (g (f x)) == f (g (f (f x))) . eta(x)

 eta^(-1)(x) . eta(y) . f (g (f x)) == f (g (f (f x)))
 eta^(-1)(x) . eta(y) . f (g x) == f (f (g x))

We want the "real" one to stay fixed on the leftmost of the compositions of fs.

goingTheOtherWayLeastActions :: ()

naturalTransformationUpToBijection :: ()

Suppose we have some natural transformation: f . g -> f . h. Is this equivalent to some natural transformation: g -> h, up to bijection?

One way, take f's identity natural transformation id_f :: f -> f. Then the left composition with (g -> h) is another natural transformation: (f . g -> f . h), and since id_f is a bijection this is as well. Q.E.D.

Another way, we have some natural transformation from (f . g -> f . h).

    eta is a natural transformation from (f . g) to (f . h).
    X :: C => eta_X :: (f . g) x -> (f . h) x
    X :: C => beta_X :: g x -> h x
    eta_Y . (f . g) fn = (f . h) fn . eta_X
      fn :: X -> Y

    eta_(beta_Y) . (f . g) fn = (f . h) fn . eta_(beta_X)

    eta_X :: f x -> f x
    beta_X :: g x -> h x

Well, this suggests that we can only go the other way when we have a bijection from (f . g) to (f . h)

Otherwise, we wouldn't be able to form bijections between (eta_X :: (f . g) x -> (f . h) x) and (beta_X :: g x -> h x).

 eta . f . beta . f' == (id :: (f . h) x -> (f . h) x) => f :: h x -> (f . g) x, f' :: (f . h) x -> g x
 beta . g . eta . g' == (id :: h x -> h x) => g :: (f . h) x -> g x, g' :: h x -> (f . g) x

 g == f . g
 g == f . h
 h == f . g
 h == f . h
 g == h

 f == f . g

 (g) must be a left identity of (f) under composition

Results:

 An injective mapping from a natural transformation :: (g -> h) to one :: (f . g -> f . h) is trivial.
 An injective mapping from a natural transformation :: (f . g -> f . h) to one :: (g -> h) exists iff (g == h) and (f == f . g), i.e. (g) must be a left identity of (f) under composition.

functorLimitNotes :: ()

Question, is the limit of some functor inhabited? e.g.

  The limit of Maybe is inhabited by:
    Nothing
    Just Nothing
    Just (Just Nothing)
    (the natural numbers)
  The limit of (x -> (a, x)) has exactly one inhabitant:
    (a, (a, (a, (a, ..))))
    (a right-infinite list)
  The limit of (x -> Either a x) has the inhabitants:
    Left a
    Right (Left a)
    Right (Right (Left a)
    (the pair (Nat, a))

what about expressing the limit of pushes? That's probably something like:

  f (g a) -> f (g (Fix f a))
  no... hmmm

functorCompositionNotes :: ()

Some notes on composing sequences of functors

f . g == f . g . f
f . g == (f . g) . f == ((f . g) . f) . f == (((f . g) . f) . f) . f
f . g == f . (f . g) == f . (f . (f . g)) == f . (f . (f . (f . g)))

a; a,b; a,b,a; a,b,a,b
b; b,a; b,a,b; b,a,b,a