Data.Flip

Documentation

newtype F1 a b c

A four-part mutually recursive data type

Notes:

Number of cycles: w.l.o.g we require the first element to be the minimum

1 minimum
* permutations of all else
(n-1)! ?

E.g. for two, there's exactly one cycle:

  Value, Container =>
    Value * Container (..), Container (Value * ..)
    Value + Container (..), Container (Value + ..)

Interesting example: unary / binary

(a -> b)
(a -> b -> c)

a -> b
a -> (b -> c)
a -> (a, b, b -> c)
b -> (b, c)
a -> (b, (b -> (c, next)))

Want:

Applicative instances for F1, F3

Comonad instances for F2, F4

Constructors

F1

Fields

getF1 :: a -> F2 a b c

newtype F2 a b c

Continuation of F1

Constructors

F2

Fields

getF2 :: (b, F3 a b c)

newtype F3 a b c

Continuation of F2

Constructors

F3

Fields

getF3 :: b -> F4 a b c

newtype F4 a b c

Continuation of F3

Constructors

F4

Fields

getF4 :: (c, F1 a b c)

freeMorphismsNotes :: ()

I believe the exact list of "free" morphisms between F's is:

1 -> 1          (we don't have an a, and F1's the only one that depends directly on a)
2 -> 1, 2, 3, 4 (see implementations)
3 -> 3          (as with F1, we don't have b..)
4 -> 1, 4       (snd gives F1, then we get stuck)

morphF11 :: F1 a b c -> F1 a b c

id

functorMorphF11 :: TestBatch

morphF21 :: F2 a b c -> F1 a b c

F2 to F4, then F4 to F1

functorMorphF21 :: TestBatch

morphF22 :: F2 a b c -> F2 a b c

id

functorMorphF22 :: TestBatch

morphF23 :: F2 a b c -> F3 a b c

Get and snd

functorMorphF23 :: TestBatch

morphF24 :: F2 a b c -> F4 a b c

Get and apply the b in F2 to the function in F3

functorMorphF24 :: TestBatch

morphF33 :: F3 a b c -> F3 a b c

id

functorMorphF33 :: TestBatch

morphF41 :: F4 a b c -> F1 a b c

Get and snd

functorMorphF41 :: TestBatch

morphF44 :: F4 a b c -> F4 a b c

id

functorMorphF44 :: TestBatch

testF1Functor :: TestBatch

testF2Functor :: TestBatch

testF3Functor :: TestBatch

testF4Functor :: TestBatch

newtype Flip f a b

Like flip

Constructors

Flip

Fields

getFlip :: f b a

withFlip :: (Flip f1 a1 b1 -> Flip f a b) -> f1 b1 a1 -> f b a

coerce

newtype Flip2 f a b c

Flip for three arguments

Constructors

Flip2

Fields

getFlip2 :: f c b a

withFlip2 :: (Flip2 f1 a1 b1 c1 -> Flip2 f a b c) -> f1 c1 b1 a1 -> f c b a

Coerce

newtype SomeIso a b

A newtype to ensure generated functions (Arbitrary) are inverses of each other

Constructors

SomeIso

Fields

getSomeIso :: (b -> a, a -> b)

testIsoFunctor :: (IsoFunctor f, EqProp (f a), Show (f a), Arbitrary (SomeIso a a), Arbitrary (f a)) => f a -> TestBatch

Test IsoFunctor

class IsoFunctor f where

A "functor" that requires an isomorphism to map over

Minimal complete definition

isomap

Methods

isomap :: (b -> a) -> (a -> b) -> f a -> f b

testIsoFunctorF1 :: TestBatch

Uses Word, testIsoFunctor

testIsoFunctorF2 :: TestBatch

Uses Word, testIsoFunctor

testIsoFunctorF3 :: TestBatch

Uses Word, testIsoFunctor

testIsoFunctorF4 :: TestBatch

Uses Word, testIsoFunctor

testBatches :: [TestBatch]

All TestBatchs

quickTestBatches :: IO ()

Run all test batches using quickBatch