Data.Functor.Result

Documentation

data Result f b c

Object, but with a sum (Either) instead of a tuple

Constructors

Result

Fields

getResult :: Either c (Effect f b c)

data Effect f b c

Action, but with a sum (Either) instead of a tuple

Constructors

Effect

Fields

runEffect :: f b (Result f b c)

cycleSum :: Result Either a b -> Either a b

Note: while this function has a simple type, if the Result cycles this functions result will be bottom