Data.Functor.Object

Documentation

data Object f b c

Both a right result and a continuation (Action)

Constructors

Object

Fields

getObject :: (c, Action f b c)

data Action f b c

An action which could contain a left result, continuation (Object)

Constructors

Action

Fields

runAction :: f b (Object f b c)

newtype Fix f

Simple fixed point newtype

Constructors

Fix

Fields

runFix :: f (Fix f)

cyclePair :: Object (,) a b -> Fix ((,) (a, b))

Cycle a pair, resulting in a fixed point

fixList :: Fix ((,) a) -> [a]

Convert a Fix to an infinite list

data Result f b c

Object, but with a sum (Either) instead of a tuple

Constructors

Result

Fields

getResult :: Either c (Effect f b c)

data Effect f b c

Action, but with a sum (Either) instead of a tuple

Constructors

Effect

Fields

runEffect :: f b (Result f b c)

fromAll :: NonEmpty (a, b) -> Action (,) a b

Given all the functional pairs to define a function, we can cyclically wrap the pairs into an Action over pairs.

Consider generalizing to other nice (* -> * -> *)'s

ish :: (a -> b -> (a', b')) -> NonEmpty a -> NonEmpty b -> Action (,) a' b'

interesting..

fstIsh :: NonEmpty a -> NonEmpty b -> Action (,) (a, b) a

Effectively define fst for an Action over pairs, using all cases

sndIsh :: NonEmpty a -> NonEmpty b -> Action (,) (a, b) b

Effectively define snd for an Action over pairs, using all cases

appIsh :: (Eq a, Functor (f t)) => Action (,) a b -> Object f t a -> Object f t b

Since there's only a finite number of a's, we can cycle through til we match, then continue on

toEither :: Flip (Result (Flip (Object f))) c b -> Either c b

Convert a flipped result to an either (sum)

toPair :: Flip (Object (Flip (Object f))) c b -> (c, b)

Convert a flipped result to a tuple (product)

unAction :: Functor (f b) => Action f b c -> f b c

Undo an Action

unObject :: Functor (f b) => Object f b c -> (f b c, c)

Undo an Object

toAction :: Functor (f b) => f b c -> Action f b c

Convert to an Action

toObject :: Functor (f b) => f b c -> c -> Object f b c

Convert to an Object

fstAction :: Functor (f (b, c)) => f (b, c) b -> Action f (b, c) b

toAction

fstObject :: Functor (f (b, c)) => f (b, c) b -> b -> Object f (b, c) b

toObject

sndAction :: Functor (f (b, c)) => f (b, c) c -> Action f (b, c) c

toAction

sndObject :: Functor (f (b, c)) => f (b, c) c -> c -> Object f (b, c) c

toObject

swapAction :: Functor (f (b, c)) => f (b, c) (c, b) -> Action f (b, c) (c, b)

toAction

swapObject :: Functor (f (b, c)) => f (b, c) (c, b) -> (c, b) -> Object f (b, c) (c, b)

toObject

mapEitherAction :: Functor (f (c -> d, Either b c)) => f (c -> d, Either b c) (Either b d) -> Action f (c -> d, Either b c) (Either b d)

toAction

mapEitherObject :: Functor (f (c -> d, Either b c)) => f (c -> d, Either b c) (Either b d) -> Either b d -> Object f (c -> d, Either b c) (Either b d)

toObject